requirements.txt
bagit~=1.7
pypandoc~=1.4
python-dotenv~=0.10
python-frontmatter~=0.4
PyYaml~=5.3.1
requests~=2.23.0
